requirements.txt
simplejson
pymysql
Flask
flask-cors
psycopg2-binary

